Print diagnostic info to stderr instead of stdout
stderr is defined in POSIX to be not only for errors, but also diagnostic
and status information. By moving the status information in `cocoapods-why`
from stdout to stderr, it no longer gets in the way when we pipe the output
of `pod why` into other scripts.

We use `warn` instead of `$stderr.puts` because RuboCop recommends it.

https://www.rubydoc.info/gems/rubocop/RuboCop/Cop/Style/StderrPuts

This is a breaking change and will require a major version bump, since
scripts may be depending on output that is currently sent to stdout.